% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_distributions.R
\name{binDirections}
\alias{binDirections}
\title{Bin observed wind directions into 16 bins}
\usage{
binDirections(v)
}
\arguments{
\item{v}{vector of observed wind directions}
}
\value{
v a vector of binned wind directions
}
\description{
\code{binDirections} returns a vector of binned wind directions
}
\details{
This function bins a vector of wind directions into 16 bins
}

