% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{buildAverages}
\alias{buildAverages}
\title{Build a dataframe with plot averages}
\usage{
buildAverages(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with averages for each plot
}
\description{
\code{buildAverages} returns dataframe with averaged data
}
\details{
This fucntion returns a dataframe with wind data averaged for each plot.
Data are for each plot averaged over all the entire dataframe.
}
\examples{
data(wind)
s.avg <- buildAverages(s)
}
