% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{buildBiasHourlyAverages}
\alias{buildBiasHourlyAverages}
\title{Build a dataframe with hourly averaged observed and predicted data}
\usage{
buildBiasHourlyAverages(df)
}
\arguments{
\item{df}{bias dataframe returned from \code{wnBuildBiasDf} subset on one model}
}
\value{
dataframe with hourly averages
}
\description{
\code{buildBiasHourlyAverages} returns dataframe with hourly averaged data
}
\details{
This fucntion returns a dataframe with hourly averages of observed
and predicted wind data. The dataframe should contain only one model for predictions.
Data are averaged over all timesteps to produce 
hourly averages for each hour of the day. This is useful to see, 
for example, what the typcial wind field looks like at 1000 LT. 
It may be most usefule to call this function after other subsetting operations have been done.
For example, you may want to first subset on speed to filter out
high-wind event cases to examine diurnal wind patterns.
}

