% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_distributions.R
\name{createLookUpTable}
\alias{createLookUpTable}
\title{Generate a lookup table}
\usage{
createLookUpTable(df, pathToRasters, rasterNames)
}
\arguments{
\item{df}{dataframe with columns 'plot', 'lat', 'lon'}

\item{pathToRasters}{path to the raster files (must be in order)}

\item{rasterNames}{vector of individual raster names}
}
\value{
dataframe
}
\description{
\code{createLookUpTable} returns a wind direction lookup table
}
\details{
This function crates a lookup table of wind directions
for use in generating wind direction distributions.
}

