% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbFetchSensorLocation}
\alias{dbFetchSensorLocation}
\title{Fetch the location (lat/lon) of one or more sensors}
\usage{
dbFetchSensorLocation(db, sensors)
}
\arguments{
\item{db}{wind database to query}

\item{sensors}{sensor(s) to extract location for}
}
\value{
dataframe with sensor id(s), lat, lon
}
\description{
\code{dbFetchSensorLocation} returns the lat/lon of one or more sensors
}
\details{
This function returns the lat/lon of one or more sensors
}
