% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbFetchWind}
\alias{dbFetchWind}
\title{Fetch all data from an SQLite database in windtools format for a specified period of time}
\usage{
dbFetchWind(db, sensors, start_time, end_time, progressBar = FALSE)
}
\arguments{
\item{db}{wind database to query}

\item{sensors}{sensors to extract info for}

\item{start_time}{format is '2011-08-15 06:00:00'}

\item{end_time}{format is '2011-08-15 6:00:00'}

\item{progressBar}{includes progress bar for shiny web interface}
}
\value{
dataframe with id, date/time, speed, gust, direction, quality, sensor quality, latitude, and longitude
}
\description{
\code{dbFetchWind} returns a dataframe from an SQLite database for a specified period of time
}
\details{
This function returns a dataframe of raw, 30-s wind data for sensor for a specified time period in the windtools package format.
}
