% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_distributions.R
\name{distributionFromTable}
\alias{distributionFromTable}
\title{Generate wind direction distribution}
\usage{
distributionFromTable(v, table)
}
\arguments{
\item{v}{vector of binned observed wind directions (number of bins has to equal the number of bins in table)}

\item{table}{the lookup table to use}
}
\value{
dataframe of wind direction distributions
}
\description{
\code{distributionFromTable} returns a wind direction distribution
}
\details{
This function generates a distribution of wind directions
based on a lookup table and vector of observed directions.
}

