% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mae}
\alias{mae}
\title{Calculate the mean absolute error}
\usage{
mae(x)
}
\arguments{
\item{x}{vector}
}
\value{
mean absolute error
}
\description{
\code{mae} returns the mean absolute error
}
\details{
This fucntion returns the mean absolute
error of a vector
}
