% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSensor}
\alias{plotSensor}
\title{Plot wind speed/direction as a time series}
\usage{
plotSensor(df, sensor, var = "speed", threshold = NULL)
}
\arguments{
\item{df}{dataframe}

\item{sensor}{name of sensor ('plot')}

\item{var}{variable to plot ('speed', 'direction', or 'both')}

\item{threshold}{threshold speed to indicate with horizontal bar; only shown when speed is plotted}
}
\value{
ggplot2 object
}
\description{
\code{plotSensor} returns speed and/or direction time series as a ggplot2 object
}
\details{
This fucntion creates a ggplot2 object of wind speed
vs. time for a single sensor.
}
\examples{
data(wind)
plotSensor(wind, 'R26')
}

