% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterToVectorMap.R
\name{rasterToVectorMap}
\alias{rasterToVectorMap}
\title{Convert speed, dir rasters to a vector map}
\usage{
rasterToVectorMap(spd, dir, lat, lon, zoom, maptype, colorscale = "discrete",
  axis_labels = TRUE)
}
\arguments{
\item{spd}{raster of wind speeds}

\item{dir}{raster of wind directions (direction wind is coming from; 0-360)}

\item{lat}{center lat for ggmap object}

\item{lon}{center lon for ggmap object}

\item{zoom}{zoom level for ggmap object (1-14)}

\item{maptype}{basemap for ggmap object (terrain, satellite, hybrid)}

\item{colorscale}{discrete or continuous (default is discrete)}

\item{axis_labels}{TRUE or FALSE (default is TRUE)}
}
\value{
ggmap object
}
\description{
\code{rasterToVectorMap} Converts spd, dir datafame to vector plot in a ggmap object
}
\details{
This function takes two rasters (speed and direction) and converts
them to a vector plot of wind overlaid on a Google Maps basemap. The
output is a ggmap object.
}

