% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_data.R
\name{readData}
\alias{readData}
\title{Read wind data}
\usage{
readData(fileName)
}
\arguments{
\item{fileName}{full path to wind data file to be read in}
}
\value{
dataframe for use by other windtools functions
}
\description{
\code{readData} reads wind data into a dataframe
}
\details{
This fucntion reads in a file containing wind data. The input
file must contain the following data in this order:
'identifier', 'lat', 'lon', 'datetime', 'obs_speed', obs_dir'
}

