\name{shinyPlotSensorSpeed}
\alias{shinyPlotSensorSpeed}
\title{Plot wind speed as a time series}
\usage{
  shinyPlotSensorSpeed(df)
}
\arguments{
  \item{df}{dataframe containing a single sensor}
}
\value{
  ggplot2 object
}
\description{
  \code{shinyPlotSensorSpeed} returns speed time series as
  a ggplot2 object
}
\details{
  This fucntion creates a ggplot2 object of wind speed vs.
  time for a single sensor.
}

