% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{uv2dir}
\alias{uv2dir}
\title{Convert u,v to wind direction}
\usage{
uv2dir(u, v)
}
\arguments{
\item{u}{}

\item{v}{}
}
\value{
wind direction
}
\description{
\code{uv2dir} returns wind direction
}
\details{
This function converts u, v to wind direction
in degrees.
}
