% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{uv2speed}
\alias{uv2speed}
\title{Convert u,v to wind speed}
\usage{
uv2speed(u, v)
}
\arguments{
\item{u}{}

\item{v}{}
}
\value{
wind speed
}
\description{
\code{uv2speed} returns wind speed
}
\details{
This function converts u, v to wind speed.
}

