\name{wnCreateBiasVectorMap}
\alias{wnCreateBiasVectorMap}
\title{Create vector maps of overlaid observed and predicted winds}
\usage{
wnCreateBiasVectorMap(df)
}
\arguments{
  \item{df}{dataframe returned from
  buildBiasHourlyAverages}
}
\value{
plotGoogleMaps object
}
\description{
\code{wnCreateBiasVectorMap} creates bubble maps of wind
prediction errors
}
\details{
Returns a vector map of observed and predcited winds.
}

