% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wn_plot_bias.R
\name{wnCreateBubbleMap}
\alias{wnCreateBubbleMap}
\title{Create bubble maps of wind prediction errors}
\usage{
wnCreateBubbleMap(df, model, var = "speed", stat = "bias", breaks = 5,
  b = NULL, c = NULL)
}
\arguments{
\item{df}{dataframe returned from wnBuildBiasDf() or wnBuildTsDf()}

\item{model}{weather model to subset on (a level in fcastName)}

\item{var}{variable to plot (speed, dir)}

\item{stat}{statistic to plot (bias, rmse)}

\item{breaks}{number of breaks to set in legend}

\item{b}{vector of numeric break points for legend (upper limits only)}
}
\value{
bubbleGoogleMaps object
}
\description{
\code{wnCreateBubbleMap} creates bubble maps of wind prediction errors
}
\details{
Returns a bubble map of wind prediction errors. WindNinja and weather
model errors are displayed in the same map.
}
