% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbFetchAll}
\alias{dbFetchAll}
\title{Fetch all data from an SQLite database for a specified time period}
\usage{
dbFetchAll(db, start_time, end_time)
}
\arguments{
\item{db}{wind database to query}

\item{start_time}{format is '2011-08-15 06:00:00'}

\item{end_time}{format is '2011-08-15 06:00:00'}
}
\value{
dataframe with id, date/time, speed, gust, direction, and quality
}
\description{
\code{dbFetchAll} returns a dataframe from an SQLite database for a specified time period
}
\details{
This fucntion returns a dataframe of raw, 30-s wind data for a specified
time period.
}

