% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbFetchAvg}
\alias{dbFetchAvg}
\title{Fetch averaged data from an SQLite database for a specified time period}
\usage{
dbFetchAvg(db, start_time, end_time, avg_time)
}
\arguments{
\item{db}{database to query}

\item{start_time}{format is '2011-08-15 06:00:00'}

\item{end_time}{format is '2011-08-15 06:00:00'}

\item{avg_time}{averaging time in minutes}
}
\value{
dataframe with id, date/time, speed, gust, direction, and quality
}
\description{
\code{dbFetchAvg} returns a dataframe from an SQLite database for a specified time period
}
\details{
This fucntion returns a dataframe of averaged data for a specified
time period. The data is averaged at the top of the hour.
Currently only connects to SQLite db.
}
