\name{sde}
\alias{sde}
\title{Calculate the standard deviation of error}
\usage{
sde(x)
}
\arguments{
  \item{x}{vector}
}
\value{
standard deviation of error
}
\description{
\code{sde} returns the standard deviation of error
}
\details{
This fucntion returns the standard deviation of error of a
vector
}

