% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wn_plot_bias.R
\name{wnPlotDirTs}
\alias{wnPlotDirTs}
\title{Make a faceted wind direction time series plot}
\usage{
wnPlotDirTs(df, color_list = NULL, xscale = NULL)
}
\arguments{
\item{df}{dataframe returned from wnBuildTsDf()}

\item{color_list}{optional list of colors to use (e.g., colors <- c("WindNinja" = "red", "NWP" = "darkblue"))}

\item{xscale}{optional scale_x_datetime() argument}
}
\value{
ggplot2 object
}
\description{
\code{wnPlotDirTs} plots wind direction time series faceted on sensor and weather model type
}
\details{
Plots the wind direction time series faceted on NWP type and model type
}

